Modify input argument to LeViT building blocks from torch module to string
Summary:
With torch module as an input argument, the config file needs a class of torch module as values. It makes exporting and importing harder. For example, torch module is not json serializable.

Instead, taking a str as input and constructing modules inside the function make it easier to export and import configs.

Reviewed By: zhanghang1989

Differential Revision: D32627803

fbshipit-source-id: 8248fd3fd7c65362c3d921951ddd704373109e16