switching to D2Go's export
Summary:
This is a pretty complex diff, not because of code but because of context. Stay with me:
 ---
Earlier, I had a hacky export function - `boltify`. Those bolt models didn't work, likely because I was passing in a batch size of 2. The error in question was an opaque type can't be typecasted error.

As a result I have to revisit `boltify`. Since this is BE week, I might as well switch to d2go's export workflow and not have any hacks in the first place :)

----

# Main Implementation to support export:
1. Introduce CoreModule that is used during export (after quantization and convert)
2. Introduce d2go warpper that supports `prepare_for_quant`, `prepare_for_quant_convert` and `prepare_for_export`
3. Set `EAGER_MODE = False` - This enables FX mode which is needed for my models?

## Explaining why we need this meta-arch:
boltification happens in a few steps:
1. quantization (fuse_fx, prepare_fx)
2. conversion (convert_fx)
3. export to script, then to bolt

Steps 1 and 2 will convert a pytorch model to a GraphModule, and lose functions other than `forward(...)` afaik. To get around that, we need the d2go wrapper to support `prepare_for_quant`, `prepare_for_quant_convert` which will modify the `model` attr inside the wrapper - so now we have the MetaArch but the actual model is a GraphModule.

The export step itself is a bit weird because it assumes that model_input is a flattened list/tuple of tensors, but thats fairly well documented and can be followed on the wiki
----

# Additional boiler plate in the Task

The d2go exporter has some quirks, especially when using LightningTask in oculus_face style.
Primarily, it calls `build_detection_train_loader` and passes in `list(cfg.DATASETS.TEST)` - which is bad for me because that config path was overwritten to be a CfgNode rather than a list of dataset names. Hence:

1. Edit orthofit task to create hive loader when `build_detection_test_loader` is called. Luckily Lightning calls `.test_loader`, which internally calls `_evaluation_loader`, so that code-path is disconnected from export code path.

----
# Quality of life changes to launch script:

1. [BE] num_cpu, num_nodes etc to follow Tong's lead
2. `stages` since I'll try to use the d2go export operator
3. When exporting, also set max_batch_count to avoid getting an infinitely looping train dataloader

Reviewed By: yanglinfang

Differential Revision: D32633897

fbshipit-source-id: b9242c46cb92c4a031bc39fbe8010ea16b0c25fd