Update the sync batch norm implementation for sync_bn
Summary:
The sync_bn batch norm has been using a non-efficient implementation NaiveSyncBatchNorm. This is because in pytorch<=1.5, nn.SyncBatchNorm had incorrect gradient when the batch size on each worker is different, so a NaiveSyncBatchNorm was implemented. This issue has been fixed in https://github.com/pytorch/pytorch/pull/36382. So we change that implementation back to the faster nn.SyncBatchNorm.

Since nn.SyncBatchNorm only has GPU implementation, for the unit tests running with single process on CPU, the original NaiveSyncBatchNorm is used.

Reviewed By: wat3rBro

Differential Revision: D35300913

fbshipit-source-id: c649f4dfc1ab5b2a0dd91d0187b072ae2de229ff