Add sanity check in create self contained model script
Summary:
The `create_self_contain_model` tools can produce a model that fails to run. This can happen if the user provide some inputs that are not valid for the specific model forward logic.
To avoid having produce an invalid output model, this diff adds some sanity check on the model and runs the forward logic after creating the self contained model.

Differential Revision: D36832923

fbshipit-source-id: f28ba644826ea2af96e8b234b075f29d3d2b1d71