Add sanity check in create self contained model script
Summary:
Pull Request resolved: https://github.com/facebookresearch/mobile-vision/pull/79

The `create_self_contain_model` tools can produce a model that fails to run. This can happen if the user provide some inputs that are not valid for the specific model forward logic.
To avoid having produce an invalid output model, this diff adds some sanity check on the model and runs the forward logic after creating the self contained model.

Reviewed By: wat3rBro

Differential Revision: D36832923

fbshipit-source-id: 129d17aa8407f99b1040fe1b6d6fbd20c2e567c2