Use TorchVision's Model Registration API
Summary:
Pull Request resolved: https://github.com/facebookresearch/mobile-vision/pull/103

TorchVision has introduced a [Model Registration API](https://pytorch.org/blog/easily-list-and-initialize-models-with-new-apis-in-torchvision/) which allows to easily list all existing models. The API is being adopted by multiple external frameworks like MMLabs, Lightning and fastai. We believe it will beneficial to add it on MobileCV so that the users of MobileVision can get access directly to the available backbones of TorchVision.

Reviewed By: wat3rBro

Differential Revision: D39374209

fbshipit-source-id: de31e15f52577ed4b62a97aa3cfe7723b22c79ab