support creating in-memory shared list from worker process
Summary:
Pull Request resolved: https://github.com/facebookresearch/mobile-vision/pull/120

A convenient helper function to share object to peer workers and child processes without using extra memory. The implementation is based on the new `multiprocessing.shared_memory` added in Python 3.8, the nice feature is that the shared objects can be created after launching worker process, and it doesn't need any preparation (eg. creating a `mp.Pipe`) from parent process. It is also modular, which only depends on native python lib and `comm` module to get the rank and do synchronization.

Reviewed By: tglik

Differential Revision: D40789062

fbshipit-source-id: 26016406673898f6265103d82df70f4e866f533b