Fix error for model with trace_type == "script"
Summary:
Pull Request resolved: https://github.com/facebookresearch/mobile-vision/pull/191

Recently, we attempted to export the IOBT model. According to the method described in the test plan of D53447647, we were able to generate the corresponding BoltNN versions of the Encoder and LSTM ptl files. Each of them can be run standalone. However, when we combined these two models and tested them together, we encountered an error (P1157420937).

After our experiments, it appears that this diff can help to resolved the issue.

Reviewed By: stephenyan1231

Differential Revision: D53483747

fbshipit-source-id: 7c46a82f9d1bb4e064972a3db8c4ed54e4dafb74