make a copy of model with trace_type script and model to be augmented with bundled inputs
Summary:

MT-Argos model export is broken after D53483747 with changes to **model_exporter** is landed.

Such broken happens when **torch.jit.save(script_model, fp, _extra_files=model_extra_files)** fails, where **script_model** contains the original boltnn model, which is successfully delegated.

It fails with a runtime error **RuntimeError: Unknown IValue type for pickling: Storage**. See full stack trace P1185839015.

The reason is **augment_model_with_bundled_inputs** make in-place changes to the input argument **model**, and inserts data with IValue of un-supported types (e.g. Storage).

To overcome this issue, only when we prepare a model with bundled model inputs and the model trace type is **script**, we make a copy of model.

This should NOT affect the exportation of the model which does not contain bundled model inputs, such as IOBT model as in D53483747.

Reviewed By: mohankumarkumar, BoomFan

Differential Revision: D53932238